+++ leshan-core/src/main/java/org/eclipse/leshan/core/node/codec/senml/LwM2mResolvedSenMLRecord.java
@@ -33,7 +33,7 @@ public class LwM2mResolvedSenMLRecord extends ResolvedSenMLRecord {
      * @throws IllegalArgumentException if path is invalid
      * @throws InvalidLwM2mPathException if path is invalid
      */
-    public LwM2mResolvedSenMLRecord(SenMLRecord unresolvedRecord, String resolvedName, Long resolvedTimestamp)
+    public LwM2mResolvedSenMLRecord(SenMLRecord unresolvedRecord, String resolvedName, Float resolvedTimestamp)
             throws InvalidLwM2mPathException {
         super(unresolvedRecord, resolvedName, resolvedTimestamp);
         this.path = new LwM2mPath(resolvedName);
